+++ iota/Vector.py
@@ -45,6 +45,30 @@ class Vector:
             return Vector(0, 0, 0)
         return Vector(self.x / mag, self.y / mag, self.z / mag)
 
+    def clamp_to_basis(self) -> "Vector":
+        x, y, z = self.x, self.y, self.z
+        a, b, c = abs(x), abs(y), abs(z)
+        
+        if b >= a and b >= c:
+            max_idx = 1
+        elif c >= a and c >= b:
+            max_idx = 2
+        else:
+            max_idx = 0
+
+
+        if max([a,b,c]) == 0:
+            return Vector(0, 0, 0)
+
+        sign = lambda v: 1 if v > 0 else -1
+        if max_idx == 0:
+            return Vector(sign(x), 0, 0)
+        elif max_idx == 1:
+            return Vector(0, sign(y), 0)
+        else:
+            return Vector(0, 0, sign(z))
+
+
     def __str__(self) -> str:
         return f"<{self.x:.2f}, {self.y:.2f}, {self.z:.2f}>"
 
@@ -85,7 +109,13 @@ class Vector:
         return self.__add__(other)
 
     def __rsub__(self, other):
-        return self.__sub__(other)
+        if isinstance(other, numberType):
+            return Vector(
+                other - self.x,
+                other - self.y,
+                other - self.z,
+            )
+        return other.__sub__(self)
 
     def __abs__(self):
         return self.magnitude()
@@ -118,8 +148,61 @@ class Vector:
                 other / self.y,
                 other / self.z,
             )
+        return other.__trudiv__(self)
     
     def __eq__(self, other):
         return (self.x == other.x and 
                 self.y == other.y and
                 self.z == other.z)
+    
+    def __pow__(self, other):
+        if isinstance(other, numberType):
+            return Vector(
+                self.x ** other,
+                self.y ** other,
+                self.z ** other,
+            )
+        elif isinstance(other, Vector):
+            return self.project_onto(other)
+        else:
+            raise TypeError
+    
+    def __rpow__(self, other):
+        if isinstance(other, numberType):
+            return Vector(
+                other ** self.x,
+                other ** self.y,
+                other ** self.z
+            )
+    
+    def __floor__(self):
+        return Vector(
+            self.x.__floor__(),
+            self.y.__floor__(),
+            self.z.__floor__()
+        )
+    
+    def __ceil__(self):
+        return Vector(
+            self.x.__ceil__(),
+            self.y.__ceil__(),
+            self.z.__ceil__()
+        )
+    
+    def __mod__(self, other):
+        if isinstance(other, numberType):
+            return Vector(
+                self.x % other,
+                self.y % other,
+                self.z % other,
+            )
+        return self.dot(other)
+    
+    def __rmod__(self, other):
+        if isinstance(other, numberType):
+            return Vector(
+                other % self.x,
+                other % self.y,
+                other % self.z,
+            )
+        return self.dot(other)